+++ corehq/apps/app_manager/static/app_manager/js/bootstrap3/nav_menu_media.js
